aider: Add `template-haskell` to the `build-depends` section in the `LambdaNat0/LambdaNat.cabal` file.
# Aider chat conversation:

USER: when I run cabal build I get """src/Interpreter.hs:7:1: error:
    Could not load module ‘Language.Haskell.TH’
    It is a member of the hidden package ‘template-haskell-2.18.0.0’.
    Perhaps you need to add ‘template-haskell’ to the build-depends in your .cabal file.
    Use -v (or `:set -v` in ghci) to see a list of the files searched for.
  |
7 | import Language.Haskell.TH (Exp)
  | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
"""

ASSISTANT: The error message indicates that the `template-haskell` package is not included in the build dependencies in your .cabal file.

To fix this, you need to add `template-haskell` to the `build-depends` section in the `LambdaNat0/LambdaNat.cabal` file.

Here is the *SEARCH/REPLACE block* for the modification:

```haskell
LambdaNat0/LambdaNat.cabal
<<<<<<< SEARCH
  build-depends:
      base >=4.7 && <5
    , array
    , containers
    , mtl
=======
  build-depends:
      base >=4.7 && <5
    , array
    , containers
    , mtl
    , template-haskell
>>>>>>> REPLACE
```

This change adds `template-haskell` to the list of dependencies, which should resolve the error you're seeing.